Move AI functions from unit_generic to Intelligent class.
